git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rmetrics@3369 5d4a7380-5c0c-0410-b446-d4a5018c0119

+++ pkg/timeSeries/R/time.R
@@ -124,7 +124,7 @@ setMethod("time" ,
           # Get Positions:
           ans <-
               if (x@format == "counts") {
-                  as.integer(x@positions)
+                  as.numeric(x@positions)
               } else {
                   timeDate(charvec = x@positions, format = x@format,
                            zone = x@FinCenter, FinCenter = x@FinCenter)
